Improve naming of FreeParameterExpression
By default, qiskit names elements of a ``ParameterVector`` are named like this 'a[<index>]'. The '[]' parentheses are not supported by the QASM 3 syntax which leads to an error when trying to run a Braket Circuit with free parameters following the '[]'-naming. Therefore, I ensured that these will be removed before the ``ParamterExpression`` is passed to create a FreeParameterExpression out of it.