Enable use of qiskit ParameterExpressions for circuit conversion: qiskit to braket (#139)
* Enable the use of qiskit ParameterVector for circuit conversion qiskit to braket

This modification allows users to convert their qiskit circuit object that was parametrized with a qiskit ParameterVector to convert it to a braket circuit.

* Reorder qiskit ParameterExpression import

* Improve naming of FreeParameterExpression

By default, qiskit names elements of a ``ParameterVector`` are named like this 'a[<index>]'. The '[]' parentheses are not supported by the QASM 3 syntax which leads to an error when trying to run a Braket Circuit with free parameters following the '[]'-naming. Therefore, I ensured that these will be removed before the ``ParamterExpression`` is passed to create a FreeParameterExpression out of it.

* add import sympify

* Ensure compatibility with QASM 3 syntax when using single-entry qiskit ``ParameterVector``

Assume this example: A one-qubit qiskit circuit with the parametrized Rx-gate as its only operation.
When using a qiskit ``ParameterVector``` of length 1 (i.e., params = ParameterVector("a", 1)), this would lead to the Parameter being called 'a[0]' which leads to a QASM 3 syntax error when trying to run the circuit.
The proposed changes ensure that the squared-parentheses '[]' are being removed in the parameter names.

* handle ParameterExpression

* support ParameterVector

* support ParameterVector with ParameterExpression

* draft: validate name

* validate name conflicts

* fix merge

* use _rename_param_vector_element once

* do not use sympify

* avoid using private member

* reuse _validate_name_conflicts

* fix docstring

---------

Co-authored-by: Jean-Christophe Jaskula <jjaskula@amazon.com>